Decompose Discord intake gate
