+++ src/services/discord/router/intake_gate.rs
@@ -1,62 +1,33 @@
-use super::super::outbound::reaction_control::{
-    ReactionControlReplyReason, send_reaction_control_reply,
-};
 use super::super::*;
 use super::intake_queue_transaction::{
-    IntakeQueueAuthorClass, IntakeQueueCommitEffects, IntakeQueueCommitOptions,
-    IntakeQueueCommitSource, IntakeQueuePendingReactionPolicy, SoftInterventionCommitRequest,
-    SoftInterventionSpec, commit_soft_intervention_transaction,
+    IntakeQueueAuthorClass, IntakeQueueCommitOptions, IntakeQueueCommitSource,
+    IntakeQueuePendingReactionPolicy, SoftInterventionCommitRequest, SoftInterventionSpec,
+    commit_soft_intervention_transaction,
 };
 
 mod busy_duplicate_notice;
+mod component_events;
+mod gate;
 mod node_override_routing;
+mod queue_effects;
+mod reaction_remove;
+mod stale_turn;
 
-pub(in crate::services::discord) fn should_process_turn_message(
-    kind: serenity::model::channel::MessageType,
-) -> bool {
-    matches!(
-        kind,
-        serenity::model::channel::MessageType::Regular
-            | serenity::model::channel::MessageType::InlineReply
-    )
-}
+pub(in crate::services::discord) use gate::should_process_turn_message;
+#[cfg(test)]
+pub(super) use queue_effects::queue_pending_reaction_for;
 
-pub(super) fn content_has_explicit_user_mention(content: &str, user_id: serenity::UserId) -> bool {
-    let raw_id = user_id.get();
-    content.contains(&format!("<@{raw_id}>")) || content.contains(&format!("<@!{raw_id}>"))
-}
-
-pub(super) fn should_skip_self_authored_turn_message(
-    author_id: serenity::UserId,
-    current_bot_id: serenity::UserId,
-) -> bool {
-    author_id == current_bot_id
-}
-
-pub(super) fn should_skip_for_missing_required_mention(
-    settings: &DiscordBotSettings,
-    effective_channel_id: serenity::ChannelId,
-    is_dm: bool,
-    content: &str,
-    bot_user_id: serenity::UserId,
-) -> bool {
-    !is_dm
-        && settings
-            .require_mention_channel_ids
-            .contains(&effective_channel_id.get())
-        && !content_has_explicit_user_mention(content, bot_user_id)
-}
-
-fn strip_leading_bot_mention(text: &str) -> String {
-    static BOT_MENTION_RE: std::sync::LazyLock<regex::Regex> = std::sync::LazyLock::new(|| {
-        regex::Regex::new(r"^<@!?\d+>\s*").expect("static bot-mention regex is valid")
-    });
-    BOT_MENTION_RE.replace(text, "").to_string()
-}
-
-fn should_start_attachment_only_turn(text: &str, saved_attachment_count: usize) -> bool {
-    saved_attachment_count > 0 && strip_leading_bot_mention(text).trim().is_empty()
-}
+use gate::{
+    bot_author_allowed_for_live_intake, should_merge_consecutive_messages,
+    should_skip_for_missing_required_mention, should_skip_human_slash_message,
+    should_skip_self_authored_turn_message, should_start_attachment_only_turn,
+    strip_leading_bot_mention,
+};
+use queue_effects::{IntakeGateQueueEffects, render_visible_queued_ack};
+use stale_turn::{
+    mailbox_has_live_active_turn_or_cleanup_stale_proof, thread_guard_force_clean_stale_thread,
+    thread_guard_should_force_clean_stale_thread,
+};
 
 async fn record_upload_history(
     shared: &std::sync::Arc<SharedData>,
@@ -94,14 +65,6 @@ async fn append_pending_uploads(
     } else {
         false
     }
-}
-
-pub(in crate::services::discord) fn bot_author_allowed_for_live_intake(
-    allowed_bot_ids: &[u64],
-    announce_bot_id: Option<u64>,
-    author_id: u64,
-) -> bool {
-    allowed_bot_ids.contains(&author_id) || announce_bot_id.is_some_and(|id| id == author_id)
 }
 
 fn session_has_usable_path(session: Option<&DiscordSession>) -> bool {
@@ -142,22 +105,6 @@ async fn can_route_unbound_direct_session(
     }
 
     has_direct_runtime_session(data, channel_id, effective_channel_id).await
-}
-
-fn should_skip_human_slash_message(
-    content: &str,
-    known_slash_commands: Option<&std::collections::HashSet<String>>,
-) -> bool {
-    if !content.starts_with('/') {
-        return false;
-    }
-
-    let command_name = content[1..].split_whitespace().next().unwrap_or("");
-    if command_name.is_empty() {
-        return false;
-    }
-
-    known_slash_commands.is_some_and(|set| set.contains(command_name))
 }
 
 async fn resolve_voice_transcript_announcement_for_intake(
@@ -246,375 +193,6 @@ async fn resolve_voice_transcript_announcement_for_intake(
         }
         tokio::time::sleep(std::time::Duration::from_millis(50)).await;
     }
-}
-
-/// Pick the queue-pending reaction emoji based on the enqueue outcome.
-/// Standalone queue head entries get `📬`; merged-into-previous entries get
-/// `➕` so users can tell merged from standalone at a glance (#1190 follow-up).
-#[cfg(test)]
-pub(super) fn queue_pending_reaction_for(outcome: super::super::MailboxEnqueueOutcome) -> char {
-    super::intake_queue_transaction::queue_pending_reaction_for(&outcome)
-}
-
-/// #3182: add the queue-pending reaction (📬 standalone / ➕ merged) and
-/// self-heal the late-add race.
-///
-/// Every `intake_gate` enqueue path adds this reaction AFTER one or more
-/// Discord `await`s (placeholder POST, queued-card render, dispatch-guard
-/// bookkeeping). During that await the active turn can finish and the
-/// queued-dispatch entrypoints (`DiscordGateway::dispatch_queued_turn` /
-/// `kickoff_idle_queues`) can dequeue THIS message and run their 📬/➕ reaction
-/// drain BEFORE the add lands — stranding the reaction on an
-/// already-promoted/processed message (the gate-path analog of the
-/// `intake_turn.rs` Surface-3 self-heal, #2036). So after the add await
-/// resolves, re-snapshot the mailbox and strip the reaction if the message is
-/// no longer queued.
-///
-/// Validity uses SOURCE MEMBERSHIP — head id OR any `source_message_ids`
-/// entry — for BOTH standalone and merged. The head-only rule is correct only
-/// for placeholder-card OWNERSHIP (one card per active turn); a merged source
-/// `B` that is still queued under a newer head `C` keeps a valid `➕`, so a
-/// head-only check would wrongly remove it. `remove_reaction_raw` is
-/// best-effort (no-op when already cleared), and only the calling provider
-/// bot's own @me reaction is removed.
-async fn add_queue_pending_reaction_self_healing(
-    ctx: &serenity::Context,
-    data: &Data,
-    channel_id: serenity::ChannelId,
-    user_msg_id: serenity::MessageId,
-    emoji: char,
-) {
-    add_reaction(&ctx.http, channel_id, user_msg_id, emoji).await;
-    let still_queued = {
-        let snapshot = mailbox_snapshot(&data.shared, channel_id).await;
-        snapshot.intervention_queue.iter().any(|intervention| {
-            intervention.message_id == user_msg_id
-                || intervention.source_message_ids.contains(&user_msg_id)
-        })
-    };
-    if !still_queued {
-        super::super::formatting::remove_reaction_raw(&ctx.http, channel_id, user_msg_id, emoji)
-            .await;
-        let ts = chrono::Local::now().format("%H:%M:%S");
-        tracing::info!(
-            "  [{ts}] 🔁 RACE: queue-pending {emoji} reacted after dequeue promotion (channel {}, msg {}); removed stale reaction",
-            channel_id,
-            user_msg_id
-        );
-    }
-}
-
-struct IntakeGateQueueEffects<'a> {
-    ctx: &'a serenity::Context,
-    data: &'a Data,
-}
-
-#[async_trait::async_trait]
-impl IntakeQueueCommitEffects for IntakeGateQueueEffects<'_> {
-    async fn enqueue_soft_intervention(
-        &mut self,
-        intervention: SoftInterventionSpec,
-    ) -> super::super::MailboxEnqueueOutcome {
-        let channel_id = intervention.channel_id;
-        mailbox_enqueue_intervention(
-            &self.data.shared,
-            &self.data.provider,
-            channel_id,
-            intervention.into_intervention(),
-        )
-        .await
-    }
-
-    async fn apply_pending_reaction(
-        &mut self,
-        channel_id: serenity::ChannelId,
-        message_id: serenity::MessageId,
-        emoji: char,
-    ) {
-        add_queue_pending_reaction_self_healing(self.ctx, self.data, channel_id, message_id, emoji)
-            .await;
-    }
-
-    fn advance_checkpoint(
-        &mut self,
-        channel_id: serenity::ChannelId,
-        message_id: serenity::MessageId,
-    ) -> u64 {
-        super::super::advance_last_message_checkpoint(
-            &self.data.shared,
-            &self.data.provider,
-            channel_id,
-            message_id,
-        )
-    }
-
-    async fn schedule_idle_kickoff(&mut self) -> usize {
-        super::super::kickoff_idle_queues(
-            self.ctx,
-            &self.data.shared,
-            &self.data.token,
-            &self.data.provider,
-        )
-        .await
-    }
-}
-
-/// #1446 Layer 2 — load the thread's persisted inflight state and report
-/// whether its `updated_at` is older than `INFLIGHT_STALENESS_THRESHOLD_SECS`.
-/// Returns `false` when no state file exists (nothing to clean) or when
-/// `updated_at` cannot be parsed (never infer staleness from missing data).
-///
-/// **Pure-classification helper only.** A stale `updated_at` is necessary
-/// but not sufficient to force-clean a live thread — `updated_at` only
-/// advances when `save_inflight_state` runs, so a healthy long Bash /
-/// large Read / slow LLM stream can legitimately go silent for minutes.
-/// `thread_guard_should_force_clean_stale_thread` adds the required
-/// secondary signal (watcher snapshot's `desynced == true`).
-#[allow(dead_code)] // #3034: #1446 Layer-2 classifier pinned by the intake-gate unit tests.
-pub(super) fn thread_guard_inflight_is_stale(
-    provider: &ProviderKind,
-    thread_id: serenity::ChannelId,
-    now_unix_secs: i64,
-) -> bool {
-    super::super::inflight::load_inflight_state(provider, thread_id.get())
-        .map(|state| {
-            super::super::inflight::inflight_state_is_stale(
-                &state,
-                now_unix_secs,
-                super::super::inflight::INFLIGHT_STALENESS_THRESHOLD_SECS,
-            )
-        })
-        .unwrap_or(false)
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum StaleActiveTurnProofClassification {
-    LiveOrUnclear,
-    RelayStalled,
-    QueueBlockedOrphan,
-    ExplicitBackgroundStatus,
-}
-
-fn classify_stale_active_turn_proof(
-    inflight: &super::super::inflight::InflightTurnState,
-    snapshot: &super::super::health::WatcherStateSnapshot,
-    now_unix_secs: i64,
-) -> StaleActiveTurnProofClassification {
-    if !super::super::inflight::inflight_state_is_stale(
-        inflight,
-        now_unix_secs,
-        super::super::inflight::INFLIGHT_STALENESS_THRESHOLD_SECS,
-    ) {
-        return StaleActiveTurnProofClassification::LiveOrUnclear;
-    }
-
-    if inflight.long_running_placeholder_active {
-        return StaleActiveTurnProofClassification::ExplicitBackgroundStatus;
-    }
-
-    if snapshot.desynced {
-        return StaleActiveTurnProofClassification::RelayStalled;
-    }
-
-    if !snapshot.inflight_state_present {
-        return StaleActiveTurnProofClassification::LiveOrUnclear;
-    }
-
-    if snapshot.mailbox_active_user_msg_id.is_some()
-        && snapshot.mailbox_active_user_msg_id != Some(inflight.user_msg_id)
-    {
-        return StaleActiveTurnProofClassification::LiveOrUnclear;
-    }
-
-    if !snapshot.attached && snapshot.tmux_session_alive != Some(true) {
-        return StaleActiveTurnProofClassification::QueueBlockedOrphan;
-    }
-
-    StaleActiveTurnProofClassification::LiveOrUnclear
-}
-
-async fn classify_channel_stale_active_turn_proof(
-    shared: &std::sync::Arc<SharedData>,
-    provider: &ProviderKind,
-    channel_id: serenity::ChannelId,
-    now_unix_secs: i64,
-) -> StaleActiveTurnProofClassification {
-    let Some(inflight) = super::super::inflight::load_inflight_state(provider, channel_id.get())
-    else {
-        return StaleActiveTurnProofClassification::LiveOrUnclear;
-    };
-    let Some(registry) = shared.health_registry.upgrade() else {
-        return StaleActiveTurnProofClassification::LiveOrUnclear;
-    };
-    let Some(snapshot) = registry
-        .snapshot_watcher_state_for_provider(provider, channel_id.get())
-        .await
-    else {
-        return StaleActiveTurnProofClassification::LiveOrUnclear;
-    };
-    classify_stale_active_turn_proof(&inflight, &snapshot, now_unix_secs)
-}
-
-/// #1446 / #1456 — full force-clean predicate. Requires stale persisted
-/// inflight state plus either:
-///   1. the watcher-state snapshot for the thread reports `desynced == true`
-///      (capture-lag, cross-owner mismatch, or live-tmux orphan with no
-///      relay heartbeat — the same conjunction the stall-watchdog uses), or
-///   2. the mailbox active-turn proof has no live owner (`attached == false`
-///      and no live tmux session), which is the queue-blocked fail-open path.
-///
-/// Without the snapshot's desync corroboration we would force-clean a
-/// healthy long-running turn whose `updated_at` simply has not advanced
-/// because no chunk hit the bridge in the last 5 minutes. The no-owner path
-/// is intentionally narrower: live tmux sessions and explicit background
-/// placeholder status are preserved. Returning `false` when the registry is
-/// unreachable is the conservative default — a missing registry happens
-/// during startup before the stall-watchdog would also be running, so
-/// deferring cleanup costs nothing.
-pub(super) async fn thread_guard_should_force_clean_stale_thread(
-    shared: &std::sync::Arc<SharedData>,
-    provider: &ProviderKind,
-    thread_id: serenity::ChannelId,
-    now_unix_secs: i64,
-) -> bool {
-    matches!(
-        classify_channel_stale_active_turn_proof(shared, provider, thread_id, now_unix_secs).await,
-        StaleActiveTurnProofClassification::RelayStalled
-            | StaleActiveTurnProofClassification::QueueBlockedOrphan
-    )
-}
-
-/// #1446 Layer 2 — perform the THREAD-GUARD's stale-thread cleanup:
-///   1. drop the parent → thread mapping so subsequent intakes do not re-
-///      trigger the guard,
-///   2. delete the thread's inflight state file (releases the durable lock
-///      whose presence convinced `mailbox_has_active_turn` the dispatch is
-///      still live),
-///   3. **clear** the thread's mailbox (cancel token + active turn anchor +
-///      pending interventions). `cancel_active_turn` alone is insufficient
-///      here — for a dead-dispatch case there is no live turn task to
-///      observe the cancel signal and call `finish_turn`, so
-///      `has_active_turn()` would stay `true` forever and the next bot
-///      message would re-enter the THREAD-GUARD's queueing branch.
-///      `mailbox_clear_channel` synchronously drops `active_request_owner`
-///      / `active_user_message_id` and reports `has_active_turn() == false`
-///      immediately on completion (see `ChannelMailboxMsg::Clear` handler
-///      in `turn_orchestrator.rs`).
-///   4. complete the bookkeeping that the missing `finish_turn` would
-///      otherwise have done: cancel the orphaned token (kill any leftover
-///      child / tmux session) and decrement `global_active`. Mirrors the
-///      `placeholder_sweeper::finalize_abandoned_mailbox` cleanup
-///      pattern so health and deferred-restart counters do not leak.
-///
-/// We never touch the parent channel's own mailbox — only the thread's.
-/// This preserves the `watcher_owns_live_relay` invariant by leaving
-/// parent-side relay state untouched.
-pub(super) async fn thread_guard_force_clean_stale_thread(
-    shared: &std::sync::Arc<SharedData>,
-    provider: &ProviderKind,
-    parent_channel_id: serenity::ChannelId,
-    thread_id: serenity::ChannelId,
-) {
-    let ts = chrono::Local::now().format("%H:%M:%S");
-    tracing::info!(
-        "  [{ts}] 🔓 THREAD-GUARD: stale inflight detected for thread {}, cleaning up and proceeding",
-        thread_id
-    );
-    shared.dispatch.thread_parents.remove(&parent_channel_id);
-    super::super::inflight::delete_inflight_state_file(provider, thread_id.get());
-    let cleared = mailbox_clear_channel(shared, provider, thread_id).await;
-    super::super::stall_recovery::finalize_orphaned_clear(
-        shared,
-        thread_id,
-        cleared.removed_token,
-        "1446_thread_guard_stale_inflight",
-    );
-}
-
-/// #2044 F7 (P3 — documentation): invariant note.
-///
-/// This recovery path delegates the `cancelled` flag + `global_active`
-/// decrement to `stall_recovery::finalize_orphaned_clear`, which has
-/// owned both side-effects since #1446 (see `stall_recovery.rs:65-89`):
-///   1. it calls `turn_bridge::cancel_active_token` on the removed
-///      token — that helper sets `token.cancelled = true` so any
-///      watchdog/voice-barge-in holding an Arc to the same token sees
-///      the cancellation;
-///   2. it calls `saturating_decrement_global_active`, mirroring what
-///      the normal `turn_bridge::mod.rs:3132-3141` and
-///      `tmux.rs:2052-2061` cleanup sites do inline.
-///
-/// Therefore this site MUST NOT also poke `cancelled` / `global_active`
-/// — doing so would double-decrement the counter (already saturating
-/// in `finalize_orphaned_clear`, but the duplicate is still a smell)
-/// and confuse audit logs. If a future change splits
-/// `finalize_orphaned_clear` or makes either side-effect conditional,
-/// this comment and the comments in the bridge/tmux peer sites must
-/// move in lockstep.
-async fn release_queue_blocked_stale_active_turn(
-    shared: &std::sync::Arc<SharedData>,
-    provider: &ProviderKind,
-    channel_id: serenity::ChannelId,
-    now_unix_secs: i64,
-) -> bool {
-    let classification =
-        classify_channel_stale_active_turn_proof(shared, provider, channel_id, now_unix_secs).await;
-    if classification != StaleActiveTurnProofClassification::QueueBlockedOrphan {
-        return false;
-    }
-
-    let ts = chrono::Local::now().format("%H:%M:%S");
-    tracing::warn!(
-        "  [{ts}] 🔓 QUEUE-GUARD: stale active-turn proof for channel {} has no live owner; releasing mailbox and proceeding",
-        channel_id
-    );
-    super::super::inflight::delete_inflight_state_file(provider, channel_id.get());
-    super::super::clear_watchdog_deadline_override(channel_id.get()).await;
-    let finish = mailbox_finish_turn(shared, provider, channel_id).await;
-    // #2044 F7: `finalize_orphaned_clear` owns both `cancelled.store(true)`
-    // and the saturating `global_active` decrement — do not duplicate them here.
-    super::super::stall_recovery::finalize_orphaned_clear(
-        shared,
-        channel_id,
-        finish.removed_token,
-        "1456_queue_blocked_stale_proof",
-    );
-    shared
-        .dispatch
-        .thread_parents
-        .retain(|_, thread_id| *thread_id != channel_id);
-    if !finish.has_pending {
-        shared.dispatch.role_overrides.remove(&channel_id);
-    }
-    true
-}
-
-async fn mailbox_has_live_active_turn_or_cleanup_stale_proof(
-    shared: &std::sync::Arc<SharedData>,
-    provider: &ProviderKind,
-    channel_id: serenity::ChannelId,
-) -> bool {
-    if !mailbox_has_active_turn(shared, channel_id).await {
-        return false;
-    }
-    if release_queue_blocked_stale_active_turn(
-        shared,
-        provider,
-        channel_id,
-        chrono::Utc::now().timestamp(),
-    )
-    .await
-    {
-        return mailbox_has_active_turn(shared, channel_id).await;
-    }
-    true
-}
-
-fn should_merge_consecutive_messages(text: &str, is_allowed_bot: bool) -> bool {
-    !is_allowed_bot
-        && !text.starts_with('!')
-        && !text.starts_with('/')
-        && !text.starts_with("DISPATCH:")
 }
 
 async fn build_reply_context(
@@ -736,823 +314,6 @@ fn format_attachment_reply_context(
     lines.join("\n")
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(super) enum RemovedControlReaction {
-    CancelQueuedTurn,
-    StopActiveTurn,
-}
-
-pub(super) fn classify_removed_control_reaction(
-    emoji: &serenity::ReactionType,
-) -> Option<RemovedControlReaction> {
-    match emoji {
-        serenity::ReactionType::Unicode(value) if value == "📬" => {
-            Some(RemovedControlReaction::CancelQueuedTurn)
-        }
-        serenity::ReactionType::Unicode(value) if value == "⏳" => {
-            Some(RemovedControlReaction::StopActiveTurn)
-        }
-        _ => None,
-    }
-}
-
-/// #3009: when a follow-up message is *merged* into the previous queue head
-/// (`MailboxEnqueueOutcome::merged == true`), the head intervention's
-/// `message_id` is rewritten to this follow-up's id while the older ids are
-/// retained as `source_message_ids` (see `turn_orchestrator::enqueue_intervention`).
-/// The previous queue head already owns a `📬 메시지 대기 중` placeholder card
-/// keyed under one of those older `source_message_ids`. Re-render and re-key
-/// that single card under the new head id instead of POSTing a brand-new card,
-/// so a burst of follow-ups collapses to ONE visible waiting placeholder.
-///
-/// Returns `true` when an existing merged-group placeholder was successfully
-/// reused (no new card needed). Returns `false` when there was no existing
-/// placeholder to reuse (e.g. the head's earlier POST failed), in which case
-/// the caller falls back to creating a fresh card.
-/// Pure decision helper for #3009 merged-placeholder reuse. Given the merged
-/// head's `source_message_ids` and a lookup into the live `queued_placeholders`
-/// map, return the `(prior_source_id, placeholder_msg_id)` of the existing
-/// waiting card that should be re-keyed under the new head id. Returns `None`
-/// when no earlier source id of the group owns a card yet (caller POSTs a fresh
-/// one). The follow-up's own id is excluded because it has no card yet.
-fn pick_reusable_merged_placeholder<F>(
-    source_message_ids: &[serenity::MessageId],
-    user_msg_id: serenity::MessageId,
-    mut lookup: F,
-) -> Option<(serenity::MessageId, serenity::MessageId)>
-where
-    F: FnMut(serenity::MessageId) -> Option<serenity::MessageId>,
-{
-    source_message_ids
-        .iter()
-        .copied()
-        .filter(|id| *id != user_msg_id)
-        .find_map(|id| lookup(id).map(|placeholder| (id, placeholder)))
-}
-
-/// Outcome of the #3009 merged-placeholder reuse attempt. Distinguishes the
-/// "reuse handled it" cases (caller must NOT post a fresh card) from the
-/// genuine "no reusable card, post a fresh one" case.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum MergedPlaceholderReuse {
-    /// Reused / no-op success: a single waiting card already represents this
-    /// merged backlog (either re-keyed here, or this follow-up is a stale
-    /// non-head source whose content is already folded into the head's card).
-    /// The caller must skip the fresh-card POST.
-    Handled,
-    /// No reusable card exists for the head and the head's own card render is
-    /// still required — the caller should POST a fresh placeholder.
-    PostFresh,
-}
-
-async fn reuse_merged_queued_placeholder(
-    ctx: &serenity::Context,
-    data: &Data,
-    channel_id: serenity::ChannelId,
-    user_msg_id: serenity::MessageId,
-) -> MergedPlaceholderReuse {
-    // Serialize the whole reuse (snapshot read → re-key → re-render) against
-    // every other `queued_placeholders` mutation / queue-exit drain on this
-    // channel. Without the lock a concurrent dispatch/queue-exit drain could
-    // remove the source mapping between our snapshot read and our re-key,
-    // leaking the card or resurrecting a stale mapping (#3009 TOCTOU guard).
-    let persist_lock = data.shared.queued_placeholders_persist_lock(channel_id);
-    let persist_guard = persist_lock.lock_owned().await;
-
-    let snapshot = mailbox_snapshot(&data.shared, channel_id).await;
-    // Locate the merged head whose source ids include this follow-up. The
-    // enqueue path rewrote its `message_id` to `user_msg_id`, so match on the
-    // source-id membership to stay robust even if the head id moved again.
-    let Some(merged_head) = snapshot
-        .intervention_queue
-        .iter()
-        .find(|intervention| intervention.source_message_ids.contains(&user_msg_id))
-    else {
-        // No longer queued (already dispatched / exited). The merge outcome
-        // told us this id folded into a prior head, so its content is not
-        // standalone — do NOT post a fresh card for it.
-        return MergedPlaceholderReuse::Handled;
-    };
-    let already_started = snapshot.active_user_message_id == Some(user_msg_id);
-    if already_started {
-        // The merged head is being dispatched; its card transition is owned by
-        // the dispatch hand-off. No fresh card needed.
-        return MergedPlaceholderReuse::Handled;
-    }
-
-    // codex round-1 P2 (identity race) + round-2 P2 (stale-source duplicate):
-    // only re-key the card when `user_msg_id` is STILL the merged head. Under
-    // bursty concurrent merges a newer follow-up can have already advanced the
-    // head past this id (rewriting `message_id`), leaving `user_msg_id` as a
-    // non-head source. Re-keying to a non-head id would let the dispatch drain
-    // delete the only queued card. AND posting a fresh card for this stale
-    // non-head source would recreate the very duplicate this path prevents —
-    // its content is already folded into the head's card. So treat a stale
-    // non-head merged source as Handled (no new card); the head's own reuse
-    // pass owns the single visible card.
-    if merged_head.message_id != user_msg_id {
-        return MergedPlaceholderReuse::Handled;
-    }
-
-    // codex round-3 P3: `enqueue_intervention` already appended this follow-up's
-    // text to the head's accumulated `text` (e.g. "A\nB"). Render the FULL
-    // merged request on the single reused card so the placeholder represents the
-    // whole pending backlog, not just the latest line (`text` is only this
-    // Discord message). Capture it before any further borrows of `snapshot`.
-    let merged_request_text = merged_head.text.clone();
-
-    // Find the existing placeholder card owned by any earlier source id of the
-    // merged group (excluding this follow-up, which has no card yet).
-    let existing =
-        pick_reusable_merged_placeholder(&merged_head.source_message_ids, user_msg_id, |id| {
-            data.shared
-                .queued
-                .queued_placeholders
-                .get(&(channel_id, id))
-                .map(|entry| *entry.value())
-        });
-    let Some((prior_source_id, placeholder_msg_id)) = existing else {
-        // We are the current head but no earlier source owns a card yet (e.g.
-        // the prior head's POST failed). A fresh card IS warranted here.
-        return MergedPlaceholderReuse::PostFresh;
-    };
-
-    // Re-key the mapping so the head id (== `user_msg_id`) owns the card the
-    // dispatch hand-off consumes, and the old source-id mapping is dropped so
-    // a later merged-drain cannot delete this still-live card. The
-    // placeholder_controller entry is keyed by the *placeholder* Discord
-    // message id, which is unchanged, so no controller re-keying is needed.
-    data.shared
-        .remove_queued_placeholder_locked(channel_id, prior_source_id);
-    data.shared
-        .insert_queued_placeholder_locked(channel_id, user_msg_id, placeholder_msg_id);
-
-    // Refresh the card body with merged request text; unchanged renders
-    // coalesce, changed ones edit.
-    let gateway = super::super::gateway::DiscordGateway::new(
-        ctx.http.clone(),
-        data.shared.clone(),
-        data.provider.clone(),
-        None,
-    );
-    let key = super::super::placeholder_controller::PlaceholderKey {
-        provider: data.provider.clone(),
-        channel_id,
-        message_id: placeholder_msg_id,
-    };
-    let queued_input = super::super::placeholder_controller::PlaceholderActiveInput {
-        reason: super::super::formatting::MonitorHandoffReason::Queued,
-        started_at_unix: chrono::Utc::now().timestamp(),
-        tool_summary: None,
-        command_summary: None,
-        reason_detail: None,
-        context_line: None,
-        request_line: Some(merged_request_text),
-        progress_line: None,
-    };
-    let outcome = data
-        .shared
-        .ui
-        .placeholder_controller
-        .ensure_queued(&gateway, key, queued_input)
-        .await;
-
-    // codex round-1 P2 (edit-failure rollback): only commit the re-key after
-    // `Edited`/`Coalesced`. On `EditFailed`/`Rejected`, restore the prior source
-    // id and let the caller fall back to fresh POST, with the persistence lock
-    // held so no drain observes the half-applied re-key.
-    if !matches!(
-        outcome,
-        super::super::placeholder_controller::PlaceholderControllerOutcome::Edited
-            | super::super::placeholder_controller::PlaceholderControllerOutcome::Coalesced
-    ) {
-        data.shared
-            .remove_queued_placeholder_locked(channel_id, user_msg_id);
-        data.shared.insert_queued_placeholder_locked(
-            channel_id,
-            prior_source_id,
-            placeholder_msg_id,
-        );
-        drop(persist_guard);
-        let ts = chrono::Local::now().format("%H:%M:%S");
-        tracing::warn!(
-            "  [{ts}] ⚠ QUEUE-ACK: merged reuse of placeholder {} for message {} in channel {} failed to render ({:?}); kept prior mapping, falling back to fresh card",
-            placeholder_msg_id,
-            user_msg_id,
-            channel_id,
-            outcome,
-        );
-        return MergedPlaceholderReuse::PostFresh;
-    }
-    drop(persist_guard);
-
-    let ts = chrono::Local::now().format("%H:%M:%S");
-    tracing::info!(
-        "  [{ts}] ➕ QUEUE-ACK: merged queued message {} into existing placeholder {} in channel {} (no new card)",
-        user_msg_id,
-        placeholder_msg_id,
-        channel_id
-    );
-    MergedPlaceholderReuse::Handled
-}
-
-/// #3082 part A — pure decision helper: of the existing queued cards on a
-/// channel (`(owner_msg_id, placeholder_msg_id)` pairs), pick the single card to
-/// re-key onto `new_head` so the channel keeps AT MOST ONE visible queued card.
-///
-/// * Never reuse a card already owned by `new_head` (it has no card yet / would
-///   be a self-reuse).
-/// * Prefer a card whose owner is still present in the live queue (`queued_ids`)
-///   — that is the canonical waiting card; fall back to any other channel card
-///   (e.g. a stale owner whose intervention already drained but whose card has
-///   not been cleaned up yet) so we coalesce rather than stack.
-/// * `None` when there is no other channel card to reuse → caller posts fresh.
-fn pick_channel_queued_placeholder(
-    channel_cards: &[(serenity::MessageId, serenity::MessageId)],
-    new_head: serenity::MessageId,
-    queued_ids: &std::collections::HashSet<serenity::MessageId>,
-) -> Option<(serenity::MessageId, serenity::MessageId)> {
-    let mut fallback: Option<(serenity::MessageId, serenity::MessageId)> = None;
-    for &(owner, placeholder) in channel_cards {
-        if owner == new_head {
-            continue;
-        }
-        if queued_ids.contains(&owner) {
-            return Some((owner, placeholder));
-        }
-        fallback.get_or_insert((owner, placeholder));
-    }
-    fallback
-}
-
-/// #3082 part A — generalize "AT MOST ONE queued card per channel/active turn".
-///
-/// `reuse_merged_queued_placeholder` only collapses *merge-eligible*
-/// follow-ups (same author, no reply boundary, within the merge rules). Messages
-/// that skip merge — a different author, a reply-boundary message, or an
-/// identical re-send outside the 10s dedup window — each used to POST their own
-/// fresh `📬` card, leaving multiple waiting cards in the timeline. This helper
-/// closes that gap: before POSTing a fresh card, reuse ANY existing queued card
-/// already owned by this channel's active turn, re-keying it onto the new head
-/// id and re-rendering it with the latest request text.
-///
-/// Returns:
-/// * `Some(true)`  — an existing card was successfully re-keyed/re-rendered onto
-///   `user_msg_id`. Caller must NOT post a fresh card.
-/// * `Some(false)` — we found a card and tried to reuse it but the re-render
-///   failed; the mapping was rolled back. Caller should fall through to the
-///   fresh POST path (which deletes/replaces as needed).
-/// * `None`        — there is no existing queued card to reuse on this channel.
-///   Caller posts a fresh card as before.
-async fn reuse_any_queued_placeholder_for_channel(
-    ctx: &serenity::Context,
-    data: &Data,
-    channel_id: serenity::ChannelId,
-    user_msg_id: serenity::MessageId,
-    text: &str,
-) -> Option<bool> {
-    // Serialize against every other `queued_placeholders` mutation on this
-    // channel (dispatch hand-off, queue-exit drain, merged reuse) — same TOCTOU
-    // guard as `reuse_merged_queued_placeholder`.
-    let persist_lock = data.shared.queued_placeholders_persist_lock(channel_id);
-    let persist_guard = persist_lock.lock_owned().await;
-
-    let snapshot = mailbox_snapshot(&data.shared, channel_id).await;
-    // Only reuse while this message is genuinely still queued behind an active
-    // turn. If it already started, the dispatch hand-off owns its card; if it is
-    // no longer queued, there is nothing to represent.
-    let still_queued = snapshot.intervention_queue.iter().any(|intervention| {
-        intervention.message_id == user_msg_id
-            || intervention.source_message_ids.contains(&user_msg_id)
-    });
-    if snapshot.active_user_message_id == Some(user_msg_id) || !still_queued {
-        return None;
-    }
-
-    // Find ANY existing queued card on this channel that does not already belong
-    // to `user_msg_id`. Prefer the card owned by a message still present in the
-    // queue (the live waiting head) so we re-key the canonical card; fall back to
-    // any other channel-scoped card otherwise. Either way the invariant is "one
-    // visible card", so reusing whichever exists is correct.
-    let queued_ids: std::collections::HashSet<serenity::MessageId> = snapshot
-        .intervention_queue
-        .iter()
-        .flat_map(|intervention| {
-            std::iter::once(intervention.message_id)
-                .chain(intervention.source_message_ids.iter().copied())
-        })
-        .collect();
-
-    let channel_cards: Vec<(serenity::MessageId, serenity::MessageId)> = data
-        .shared
-        .queued
-        .queued_placeholders
-        .iter()
-        .filter_map(|entry| {
-            let (ch, owner) = *entry.key();
-            (ch == channel_id).then_some((owner, *entry.value()))
-        })
-        .collect();
-
-    let Some((prior_owner, placeholder_msg_id)) =
-        pick_channel_queued_placeholder(&channel_cards, user_msg_id, &queued_ids)
-    else {
-        return None;
-    };
-
-    // Re-key the single card onto this message id so the dispatch hand-off /
-    // queue-exit drain track exactly one card per active turn.
-    data.shared
-        .remove_queued_placeholder_locked(channel_id, prior_owner);
-    data.shared
-        .insert_queued_placeholder_locked(channel_id, user_msg_id, placeholder_msg_id);
-
-    let gateway = super::super::gateway::DiscordGateway::new(
-        ctx.http.clone(),
-        data.shared.clone(),
-        data.provider.clone(),
-        None,
-    );
-    let key = super::super::placeholder_controller::PlaceholderKey {
-        provider: data.provider.clone(),
-        channel_id,
-        message_id: placeholder_msg_id,
-    };
-    let queued_input = super::super::placeholder_controller::PlaceholderActiveInput {
-        reason: super::super::formatting::MonitorHandoffReason::Queued,
-        started_at_unix: chrono::Utc::now().timestamp(),
-        tool_summary: None,
-        command_summary: None,
-        reason_detail: None,
-        context_line: None,
-        request_line: Some(text.to_string()),
-        progress_line: None,
-    };
-    let outcome = data
-        .shared
-        .ui
-        .placeholder_controller
-        .ensure_queued(&gateway, key, queued_input)
-        .await;
-
-    if !matches!(
-        outcome,
-        super::super::placeholder_controller::PlaceholderControllerOutcome::Edited
-            | super::super::placeholder_controller::PlaceholderControllerOutcome::Coalesced
-    ) {
-        // Render failed — roll back the re-key so the prior owner keeps the card,
-        // and let the caller post a fresh one (which handles deletion/cleanup).
-        data.shared
-            .remove_queued_placeholder_locked(channel_id, user_msg_id);
-        data.shared
-            .insert_queued_placeholder_locked(channel_id, prior_owner, placeholder_msg_id);
-        drop(persist_guard);
-        let ts = chrono::Local::now().format("%H:%M:%S");
-        tracing::warn!(
-            "  [{ts}] ⚠ QUEUE-ACK: channel-wide reuse of placeholder {} for message {} in channel {} failed to render ({:?}); kept prior mapping, falling back to fresh card",
-            placeholder_msg_id,
-            user_msg_id,
-            channel_id,
-            outcome,
-        );
-        return Some(false);
-    }
-    drop(persist_guard);
-
-    let ts = chrono::Local::now().format("%H:%M:%S");
-    tracing::info!(
-        "  [{ts}] ➕ QUEUE-ACK: coalesced queued message {} into existing channel placeholder {} in channel {} (one card per active turn)",
-        user_msg_id,
-        placeholder_msg_id,
-        channel_id
-    );
-    Some(true)
-}
-
-async fn render_visible_queued_ack(
-    ctx: &serenity::Context,
-    data: &Data,
-    channel_id: serenity::ChannelId,
-    user_msg_id: serenity::MessageId,
-    text: &str,
-    merged: bool,
-) -> bool {
-    // #3009: a merged follow-up reuses the existing waiting placeholder instead
-    // of stacking a duplicate card. Only fall through to a fresh POST when this
-    // message is the current merged head AND no prior card exists to reuse
-    // (e.g. the head's earlier POST failed). A stale non-head merged source is
-    // already represented by the head's single card, so it is `Handled` (no new
-    // card) — posting one would recreate the duplicate this path prevents.
-    if merged {
-        match reuse_merged_queued_placeholder(ctx, data, channel_id, user_msg_id).await {
-            MergedPlaceholderReuse::Handled => return true,
-            MergedPlaceholderReuse::PostFresh => {}
-        }
-    }
-    // #3082 part A: even when this message did NOT merge (different author, a
-    // reply-boundary message, or an identical re-send outside the dedup window),
-    // enforce a single visible queued card per active turn by reusing any card
-    // already owned by this channel's active turn instead of POSTing a second
-    // one. `Some(true)` => handled (no fresh card); `Some(false)`/`None` => fall
-    // through to the fresh POST below.
-    if let Some(true) =
-        reuse_any_queued_placeholder_for_channel(ctx, data, channel_id, user_msg_id, text).await
-    {
-        return true;
-    }
-    let post_result = super::super::gateway::send_intake_placeholder(
-        ctx.http.clone(),
-        data.shared.clone(),
-        channel_id,
-        None,
-        // #3082 P2-3: this IS the queued-turn "📬" notice — wait behind any
-        // in-flight multi-chunk answer flush so the card lands as a trailing
-        // notice, never interleaved between answer chunks.
-        true,
-    )
-    .await;
-    let placeholder_msg_id = match post_result {
-        Ok(msg_id) => msg_id,
-        Err(error) => {
-            let ts = chrono::Local::now().format("%H:%M:%S");
-            tracing::warn!(
-                "  [{ts}] ⚠ QUEUE-ACK: placeholder POST failed for queued message {} in channel {}: {}",
-                user_msg_id,
-                channel_id,
-                error
-            );
-            // #1984 (codex C — observation): record the failure without
-            // changing user-visible behaviour. The user message is already
-            // enqueued in the mailbox at this point — only the visible
-            // "queued" card rendering is missing — so the recovery label
-            // reflects that the message itself is not lost.
-            crate::services::observability::emit_intake_placeholder_post_failed(
-                data.provider.as_str(),
-                channel_id.get(),
-                Some(user_msg_id.get()),
-                "queue_ack_visible",
-                "queued_card_skipped",
-                &error.to_string(),
-            );
-            // #2044 F13: best-effort user-visible fallback. Without
-            // this, a Discord 5xx on the queued-card POST left users
-            // with "no card / no reaction yet / message silently
-            // queued" until the per-message 📬 reaction landed later
-            // — surface a short text reply so the user knows their
-            // message is in the queue even when the card path
-            // failed. Reply is rate-limited and best-effort.
-            send_reaction_control_reply(
-                ctx,
-                &data.shared,
-                channel_id,
-                user_msg_id,
-                ReactionControlReplyReason::QueuedCardPostFailed,
-                "📬 큐에 추가됨 — 카드 표시는 실패했지만 메시지는 큐잉되었습니다.",
-            )
-            .await;
-            return false;
-        }
-    };
-
-    let persist_lock = data.shared.queued_placeholders_persist_lock(channel_id);
-    let persist_guard = persist_lock.lock_owned().await;
-    let snapshot = mailbox_snapshot(&data.shared, channel_id).await;
-    // codex round-3 P2: for a merged fallback (reuse returned `PostFresh`),
-    // require this id to STILL be the queued head before committing a fresh
-    // card. Re-snapshotting under the persist lock closes the burst race where
-    // a newer follow-up merged (advancing the head past `user_msg_id`) while
-    // our `send_intake_placeholder` POST was in flight — without this, a stale
-    // non-head source could keep its own card and recreate the #3009 duplicate.
-    // Non-merged (standalone) messages keep the original source-id membership
-    // check: they are their own head, so the two conditions coincide.
-    let still_queued = snapshot.intervention_queue.iter().any(|intervention| {
-        if merged {
-            intervention.message_id == user_msg_id
-        } else {
-            intervention.message_id == user_msg_id
-                || intervention.source_message_ids.contains(&user_msg_id)
-        }
-    });
-    let already_started = snapshot.active_user_message_id == Some(user_msg_id);
-    if already_started || !still_queued {
-        drop(persist_guard);
-        // #2044 F13: route Discord 5xx/404 on delete_message into the
-        // deferred clear queue instead of silently swallowing them.
-        // Without this, a stale "📬 메시지 대기 중" card could remain
-        // visible after the turn already started or the queue moved on.
-        if let Err(error) = channel_id
-            .delete_message(&ctx.http, placeholder_msg_id)
-            .await
-        {
-            tracing::warn!(
-                "  [{ts}] ⚠ QUEUE-ACK: delete placeholder {} in channel {} failed ({error}); deferring cleanup",
-                placeholder_msg_id,
-                channel_id,
-                ts = chrono::Local::now().format("%H:%M:%S"),
-            );
-            data.shared
-                .add_pending_queue_exit_placeholder_clear_one(
-                    channel_id,
-                    user_msg_id,
-                    placeholder_msg_id,
-                )
-                .await;
-        }
-        let ts = chrono::Local::now().format("%H:%M:%S");
-        tracing::info!(
-            "  [{ts}] 🔁 QUEUE-ACK: queued message {} in channel {} no longer needs a queued card",
-            user_msg_id,
-            channel_id
-        );
-        return false;
-    }
-
-    data.shared
-        .insert_queued_placeholder_locked(channel_id, user_msg_id, placeholder_msg_id);
-    let gateway = super::super::gateway::DiscordGateway::new(
-        ctx.http.clone(),
-        data.shared.clone(),
-        data.provider.clone(),
-        None,
-    );
-    let key = super::super::placeholder_controller::PlaceholderKey {
-        provider: data.provider.clone(),
-        channel_id,
-        message_id: placeholder_msg_id,
-    };
-    let queued_input = super::super::placeholder_controller::PlaceholderActiveInput {
-        reason: super::super::formatting::MonitorHandoffReason::Queued,
-        started_at_unix: chrono::Utc::now().timestamp(),
-        tool_summary: None,
-        command_summary: None,
-        reason_detail: None,
-        context_line: None,
-        request_line: Some(text.to_string()),
-        progress_line: None,
-    };
-    let outcome = data
-        .shared
-        .ui
-        .placeholder_controller
-        .ensure_queued(&gateway, key, queued_input)
-        .await;
-    if matches!(
-        outcome,
-        super::super::placeholder_controller::PlaceholderControllerOutcome::Edited
-            | super::super::placeholder_controller::PlaceholderControllerOutcome::Coalesced
-    ) {
-        drop(persist_guard);
-        let ts = chrono::Local::now().format("%H:%M:%S");
-        tracing::info!(
-            "  [{ts}] 📬 QUEUE-ACK: rendered queued placeholder for message {} in channel {}",
-            user_msg_id,
-            channel_id
-        );
-        return true;
-    }
-
-    data.shared
-        .remove_queued_placeholder_locked(channel_id, user_msg_id);
-    drop(persist_guard);
-    // #2044 F13: same deferred-cleanup fallback as the early-return
-    // branch above — a delete_message error here is logged + enqueued
-    // for the next `drain_pending_queue_exit_placeholder_clears` pass
-    // instead of being silently discarded.
-    if let Err(error) = channel_id
-        .delete_message(&ctx.http, placeholder_msg_id)
-        .await
-    {
-        tracing::warn!(
-            "  [{ts}] ⚠ QUEUE-ACK: delete placeholder {} in channel {} failed after render miss ({error}); deferring cleanup",
-            placeholder_msg_id,
-            channel_id,
-            ts = chrono::Local::now().format("%H:%M:%S"),
-        );
-        data.shared
-            .add_pending_queue_exit_placeholder_clear_one(
-                channel_id,
-                user_msg_id,
-                placeholder_msg_id,
-            )
-            .await;
-    }
-    let ts = chrono::Local::now().format("%H:%M:%S");
-    tracing::warn!(
-        "  [{ts}] ⚠ QUEUE-ACK: queued placeholder render failed for message {} in channel {}; deleted placeholder",
-        user_msg_id,
-        channel_id
-    );
-    false
-}
-
-async fn handle_reaction_remove(
-    ctx: &serenity::Context,
-    removed_reaction: &serenity::Reaction,
-    data: &Data,
-) -> Result<(), Error> {
-    let Some(action) = classify_removed_control_reaction(&removed_reaction.emoji) else {
-        return Ok(());
-    };
-    let Some(user_id) = removed_reaction.user_id else {
-        return Ok(());
-    };
-    if user_id == ctx.cache.current_user().id {
-        return Ok(());
-    }
-
-    // Ignore reactions removed by ANY bot — only human users should be able
-    // to cancel turns via reaction removal. Bots (announce/notify) remove
-    // reactions during dispatch status sync, which races with active turns
-    // in the same thread (#670).
-    // If user is not in cache, fetch from API before deciding.
-    let cache_result = ctx.cache.user(user_id).map(|u| u.bot);
-    let is_bot = match cache_result {
-        Some(bot) => bot,
-        None => {
-            // Cache miss — fetch from Discord API to determine bot status.
-            //
-            // #2044 F4: on API failure (transient 5xx, timeout, or
-            // restart-warmup with empty cache) the previous fallback
-            // treated the user as a bot, which silently dropped
-            // legitimate user stop/cancel reactions. The reaction event
-            // already carried the user_id and we've passed the bot
-            // self-check above, so fail-open + warn is the safer
-            // default: a missed bot-self event will be re-filtered on
-            // the next cache fill, but a missed human stop is
-            // user-visible.
-            match ctx.http.get_user(user_id).await {
-                Ok(user) => user.bot,
-                Err(err) => {
-                    tracing::warn!(
-                        "  [reaction-remove] failed to fetch user {} from API: {err}; defaulting to non-bot (#2044 F4 fail-open)",
-                        user_id
-                    );
-                    false
-                }
-            }
-        }
-    };
-    if is_bot {
-        return Ok(());
-    }
-
-    let channel_id = removed_reaction.channel_id;
-    let settings_snapshot = { data.shared.settings.read().await.clone() };
-    if validate_live_channel_routing_with_dm_hint(
-        ctx,
-        &data.provider,
-        &settings_snapshot,
-        channel_id,
-        Some(removed_reaction.guild_id.is_none()),
-    )
-    .await
-    .is_err()
-    {
-        return Ok(());
-    }
-
-    // Reaction-removal controls must never imprint owner state.
-    // Only already-authorized users may trigger queue cancel / turn stop.
-    if !super::super::discord_io::user_is_authorized(&settings_snapshot, user_id.get()) {
-        return Ok(());
-    }
-
-    match action {
-        RemovedControlReaction::CancelQueuedTurn => {
-            // The 🚫 reaction added by `apply_queue_exit_feedback`
-            // (see `mod.rs:queue_exit_feedback_emoji`) is the only feedback
-            // we surface here — no extra reply, per operator preference.
-            let removed = mailbox_cancel_soft_intervention(
-                &data.shared,
-                &data.provider,
-                channel_id,
-                removed_reaction.message_id,
-            )
-            .await;
-            if removed.is_some() {
-                super::super::advance_last_message_checkpoint(
-                    &data.shared,
-                    &data.provider,
-                    channel_id,
-                    removed_reaction.message_id,
-                );
-                let ts = chrono::Local::now().format("%H:%M:%S");
-                tracing::info!(
-                    "  [{ts}] 📭 QUEUE-CANCEL: removed queued message {} in channel {} via reaction removal",
-                    removed_reaction.message_id,
-                    channel_id
-                );
-            }
-        }
-        RemovedControlReaction::StopActiveTurn => {
-            // #441: flows through cancel_text_stop_token_mailbox (mailbox_cancel_active_turn)
-            // → cancel_active_token → token.cancelled triggers turn_bridge loop exit
-            // → mailbox_finish_turn canonical cleanup
-            //
-            // #2044 F1 (TOCTOU): snapshot the cancel-token together with
-            // the active user_message_id. Between this snapshot and the
-            // cancel await, the mailbox actor may finish the current turn
-            // and start a new one for a queued message — using the
-            // snapshotted token identity via
-            // `cancel_text_stop_token_mailbox_if_current` ensures we only
-            // cancel if the mailbox is still on the same turn we just
-            // observed. The inflight-file fallback intentionally does NOT
-            // carry a token (it's only consulted when the mailbox snapshot
-            // lacks an active turn), so in that branch we fall back to the
-            // legacy unchecked cancel as before.
-            let snapshot = mailbox_snapshot(&data.shared, channel_id).await;
-            let (active_message_id, expected_token) = match snapshot.active_user_message_id {
-                Some(active_id) => (Some(active_id), snapshot.cancel_token.clone()),
-                None => {
-                    // user_msg_id == 0 (e.g. a TUI-direct turn) anchors no
-                    // Discord message that could carry a reaction, so it yields
-                    // None (never matches `removed_reaction.message_id`);
-                    // `MessageId::new(0)` would panic.
-                    let inflight_id = super::super::inflight::load_inflight_state(
-                        &data.provider,
-                        channel_id.get(),
-                    )
-                    .and_then(|state| {
-                        super::super::inflight::optional_message_id(state.user_msg_id)
-                    });
-                    (inflight_id, None)
-                }
-            };
-            if active_message_id != Some(removed_reaction.message_id) {
-                return Ok(());
-            }
-
-            let stop_lookup = if let Some(expected) = expected_token {
-                super::message_handler::cancel_text_stop_token_mailbox_if_current(
-                    &data.shared,
-                    &data.provider,
-                    channel_id,
-                    expected,
-                    "reaction remove ⏳ (if_current)",
-                )
-                .await
-            } else {
-                super::message_handler::cancel_text_stop_token_mailbox(
-                    &data.shared,
-                    &data.provider,
-                    channel_id,
-                )
-                .await
-            };
-            match stop_lookup {
-                super::message_handler::TextStopLookup::Stop(token) => {
-                    // #1218: stop_active_turn sends the provider abort key
-                    // (C-c) FIRST so the CLI sees the interrupt while its
-                    // tmux pane is still alive, then flips the cooperative
-                    // flag and SIGKILLs the wrapper. The previous order
-                    // killed the tmux-wrapper first — tearing down the
-                    // tmux session — which made the follow-up send-keys
-                    // fail with "can't find pane". For Codex/Qwen TUIs and
-                    // resumed runs (`child_pid = None`) the C-c is the
-                    // only mechanism that actually stops the provider.
-                    super::super::turn_bridge::stop_active_turn(
-                        &data.provider,
-                        &token,
-                        super::super::turn_bridge::TmuxCleanupPolicy::PreserveSession,
-                        "reaction remove ⏳",
-                    )
-                    .await;
-                    let ts = chrono::Local::now().format("%H:%M:%S");
-                    tracing::info!(
-                        "  [{ts}] 🛑 TURN-STOP: cancelled active turn for message {} in channel {} via reaction removal",
-                        removed_reaction.message_id,
-                        channel_id
-                    );
-                    // #3650: no separate notify-bot stop message — the in-place
-                    // `[Stopped]` edit on the assistant message and the 🛑
-                    // reaction already cover the stop signal.
-                }
-                super::message_handler::TextStopLookup::AlreadyStopping => {
-                    send_reaction_control_reply(
-                        ctx,
-                        &data.shared,
-                        channel_id,
-                        removed_reaction.message_id,
-                        ReactionControlReplyReason::AlreadyStopping,
-                        "Already stopping...",
-                    )
-                    .await;
-                }
-                super::message_handler::TextStopLookup::NoActiveTurn => {}
-            }
-        }
-    }
-
-    Ok(())
-}
-
-pub(super) fn is_model_picker_component_custom_id(
-    custom_id: &str,
-    fallback_channel_id: serenity::ChannelId,
-) -> bool {
-    super::super::commands::parse_model_picker_custom_id(custom_id, fallback_channel_id).is_some()
-}
-
 pub(in crate::services::discord) async fn handle_event(
     ctx: &serenity::Context,
     event: &serenity::FullEvent,
@@ -1562,28 +323,12 @@ pub(in crate::services::discord) async fn handle_event(
     match event {
         serenity::FullEvent::InteractionCreate { interaction } => {
             if let Some(component) = interaction.as_message_component() {
-                if is_model_picker_component_custom_id(
-                    &component.data.custom_id,
-                    component.channel_id,
-                ) {
-                    let settings_snapshot = { data.shared.settings.read().await.clone() };
-                    if !super::super::provider_handles_channel(
-                        ctx,
-                        &data.provider,
-                        &settings_snapshot,
-                        component.channel_id,
-                    )
-                    .await
-                    {
-                        let ts = chrono::Local::now().format("%H:%M:%S");
-                        tracing::info!(
-                            "  [{ts}] ⏭ COMPONENT-GUARD: skipping model picker in channel {} for provider {}",
-                            component.channel_id,
-                            data.provider.as_str()
-                        );
-                        return Ok(());
-                    }
-                    return handle_model_picker_interaction(ctx, component, data).await;
+                if component_events::handle_model_picker_component_if_applicable(
+                    ctx, component, data,
+                )
+                .await?
+                {
+                    return Ok(());
                 }
                 if super::super::idle_recap_interaction::is_idle_recap_custom_id(
                     &component.data.custom_id,
@@ -1633,7 +378,7 @@ pub(in crate::services::discord) async fn handle_event(
             }
         }
         serenity::FullEvent::ReactionRemove { removed_reaction } => {
-            handle_reaction_remove(ctx, removed_reaction, data).await?;
+            reaction_remove::handle_reaction_remove(ctx, removed_reaction, data).await?;
         }
         serenity::FullEvent::Message { new_message } => {
             // ── Universal message-ID dedup ─────────────────────────────
@@ -2840,494 +1585,10 @@ pub(in crate::services::discord) async fn handle_event(
     Ok(())
 }
 
-use super::super::model_picker_interaction::handle_model_picker_interaction;
-
-/// #1446 Layer 2 — `thread_guard_inflight_is_stale` reads inflight files
-/// via the runtime root override, so we keep the always-on slice that
-/// only exercises the read+staleness classification (no `SharedData`
-/// construction). The `thread_guard_force_clean_stale_thread` integration
-/// test that drives mailbox cancel / dispatch_thread_parents removal is
-/// still not in the default suite because it depends on `TestHealthHarness`.
-#[cfg(test)]
-mod thread_guard_stale_pure_tests {
-    use super::*;
-    use chrono::TimeZone;
-    use poise::serenity_prelude::ChannelId;
-
-    /// Anchor `now` and produce a stale `updated_at` literal using the
-    /// production `now_string` encoding.
-    fn local_at_offset(now_unix: i64, offset_secs: i64) -> String {
-        chrono::Local
-            .timestamp_opt(now_unix + offset_secs, 0)
-            .single()
-            .expect("valid local time")
-            .format("%Y-%m-%d %H:%M:%S")
-            .to_string()
-    }
-
-    fn seed_inflight_with_updated_at(provider: &ProviderKind, channel_id: u64, updated_at: &str) {
-        let mut state = super::super::super::inflight::InflightTurnState::new(
-            provider.clone(),
-            channel_id,
-            Some("test-thread-guard".to_string()),
-            42,
-            8_001,
-            8_002,
-            "test-input".to_string(),
-            Some("test-session".to_string()),
-            Some("test-tmux".to_string()),
-            None,
-            None,
-            0,
-        );
-        state.updated_at = updated_at.to_string();
-        state.started_at = updated_at.to_string();
-        let root = super::super::super::inflight::inflight_runtime_root()
-            .expect("inflight runtime root must be available under test override");
-        let provider_dir = root.join(provider.as_str());
-        std::fs::create_dir_all(&provider_dir).expect("create provider dir");
-        let path = provider_dir.join(format!("{channel_id}.json"));
-        let json = serde_json::to_string_pretty(&state).expect("serialize seeded inflight");
-        std::fs::write(&path, json).expect("write seeded inflight");
-    }
-
-    fn inflight_with_updated_at(
-        provider: &ProviderKind,
-        channel_id: u64,
-        updated_at: &str,
-    ) -> super::super::super::inflight::InflightTurnState {
-        let mut state = super::super::super::inflight::InflightTurnState::new(
-            provider.clone(),
-            channel_id,
-            Some("test-thread-guard".to_string()),
-            42,
-            8_001,
-            8_002,
-            "test-input".to_string(),
-            Some("test-session".to_string()),
-            Some("stale-proof-tmux".to_string()),
-            None,
-            None,
-            0,
-        );
-        state.updated_at = updated_at.to_string();
-        state.started_at = updated_at.to_string();
-        state
-    }
-
-    fn watcher_snapshot(
-        provider: &ProviderKind,
-        channel_id: u64,
-        user_msg_id: u64,
-        attached: bool,
-        tmux_session_alive: Option<bool>,
-        desynced: bool,
-    ) -> super::super::super::health::WatcherStateSnapshot {
-        let relay_health = super::super::super::relay_health::RelayHealthSnapshot {
-            provider: provider.as_str().to_string(),
-            channel_id,
-            active_turn: super::super::super::relay_health::RelayActiveTurn::Foreground,
-            tmux_session: Some("stale-proof-tmux".to_string()),
-            tmux_alive: tmux_session_alive,
-            watcher_attached: attached,
-            watcher_attached_stale: false,
-            watcher_owner_channel_id: attached.then_some(channel_id),
-            watcher_owns_live_relay: false,
-            bridge_inflight_present: true,
-            bridge_current_msg_id: Some(8_002),
-            mailbox_has_cancel_token: true,
-            mailbox_active_user_msg_id: Some(user_msg_id),
-            queue_depth: 0,
-            pending_discord_callback_msg_id: Some(8_002),
-            pending_thread_proof: false,
-            parent_channel_id: None,
-            thread_channel_id: None,
-            last_relay_ts_ms: None,
-            last_outbound_activity_ms: None,
-            last_capture_offset: None,
-            last_relay_offset: 0,
-            unread_bytes: None,
-            desynced,
-            stale_thread_proof: false,
-        };
-        let relay_stall_state =
-            super::super::super::relay_health::RelayStallClassifier::classify(&relay_health);
-        super::super::super::health::WatcherStateSnapshot {
-            provider: provider.as_str().to_string(),
-            attached,
-            tmux_session: Some("stale-proof-tmux".to_string()),
-            watcher_owner_channel_id: attached.then_some(channel_id),
-            last_relay_offset: 0,
-            inflight_state_present: true,
-            last_relay_ts_ms: 0,
-            last_capture_offset: None,
-            unread_bytes: None,
-            desynced,
-            reconnect_count: 0,
-            inflight_started_at: None,
-            inflight_updated_at: None,
-            inflight_user_msg_id: Some(user_msg_id),
-            inflight_current_msg_id: Some(8_002),
-            tmux_session_alive,
-            has_pending_queue: false,
-            mailbox_active_user_msg_id: Some(user_msg_id),
-            inflight_terminal_delivery_committed: false,
-            relay_stall_state,
-            relay_health,
-        }
-    }
-
-    /// Scoped env-var override for `AGENTDESK_ROOT_DIR`. Restores the
-    /// previous value (or removes the var) on drop. Used so the always-on
-    /// test does not leak state into adjacent test runs that may also rely
-    /// on the runtime root.
-    ///
-    /// #2444 follow-up: acquires `shared_test_env_lock()` so this writer
-    /// serializes with every other AGENTDESK_ROOT_DIR mutator in the test
-    /// binary (claude_tui::hook_relay, credential, integration tests etc),
-    /// closing the cross-module env race that survived the wave-D fix.
-    struct EnvRootGuard {
-        previous: Option<std::ffi::OsString>,
-        _lock: std::sync::MutexGuard<'static, ()>,
-    }
-    impl EnvRootGuard {
-        fn set(path: &std::path::Path) -> Self {
-            let lock = crate::config::shared_test_env_lock()
-                .lock()
-                .unwrap_or_else(|poison| poison.into_inner());
-            let previous = std::env::var_os("AGENTDESK_ROOT_DIR");
-            unsafe { std::env::set_var("AGENTDESK_ROOT_DIR", path) };
-            Self {
-                previous,
-                _lock: lock,
-            }
-        }
-    }
-    impl Drop for EnvRootGuard {
-        fn drop(&mut self) {
-            match self.previous.take() {
-                Some(value) => unsafe { std::env::set_var("AGENTDESK_ROOT_DIR", value) },
-                None => unsafe { std::env::remove_var("AGENTDESK_ROOT_DIR") },
-            }
-        }
-    }
-
-    /// `thread_guard_inflight_is_stale` must:
-    ///   1. report `true` for a stale on-disk inflight,
-    ///   2. report `false` for a fresh on-disk inflight,
-    ///   3. report `false` when the inflight file does not exist (nothing
-    ///      to clean — never cleanup a thread we know nothing about).
-    #[tokio::test]
-    async fn thread_guard_inflight_is_stale_classifies_disk_state() {
-        let temp = tempfile::tempdir().expect("create temp runtime root");
-        let _guard = EnvRootGuard::set(temp.path());
-
-        let provider = ProviderKind::Codex;
-        let now_unix = chrono::Utc::now().timestamp();
-
-        // Missing inflight → not stale.
-        assert!(
-            !super::thread_guard_inflight_is_stale(
-                &provider,
-                ChannelId::new(900_000_000_000_900),
-                now_unix
-            ),
-            "missing inflight must NOT be classified as stale"
-        );
-
-        // Stale inflight → stale.
-        let stale_channel = 900_000_000_000_901u64;
-        let stale_at = local_at_offset(
-            now_unix,
-            -(super::super::super::inflight::INFLIGHT_STALENESS_THRESHOLD_SECS as i64) - 5,
-        );
-        seed_inflight_with_updated_at(&provider, stale_channel, &stale_at);
-        assert!(
-            super::thread_guard_inflight_is_stale(
-                &provider,
-                ChannelId::new(stale_channel),
-                now_unix
-            ),
-            "stale inflight (updated_at={stale_at}) must be classified as stale"
-        );
-
-        // Fresh inflight → not stale.
-        let fresh_channel = 900_000_000_000_902u64;
-        let fresh_at = local_at_offset(now_unix, -5);
-        seed_inflight_with_updated_at(&provider, fresh_channel, &fresh_at);
-        assert!(
-            !super::thread_guard_inflight_is_stale(
-                &provider,
-                ChannelId::new(fresh_channel),
-                now_unix
-            ),
-            "fresh inflight (updated_at={fresh_at}) must NOT be classified as stale"
-        );
-    }
-
-    /// #1456: a stale active-turn proof with no attached watcher and no live
-    /// tmux owner must be classified as queue-blocked orphan state. The intake
-    /// gate uses this to release the mailbox before the new user message takes
-    /// the normal streaming path instead of being queued forever.
-    #[test]
-    fn stale_active_turn_proof_classifies_no_owner_as_queue_blocked_orphan() {
-        let provider = ProviderKind::Codex;
-        let channel_id = 900_000_000_000_910u64;
-        let now_unix = chrono::Utc::now().timestamp();
-        let stale_at = local_at_offset(
-            now_unix,
-            -(super::super::super::inflight::INFLIGHT_STALENESS_THRESHOLD_SECS as i64) - 5,
-        );
-        let inflight = inflight_with_updated_at(&provider, channel_id, &stale_at);
-        let snapshot = watcher_snapshot(
-            &provider,
-            channel_id,
-            inflight.user_msg_id,
-            false,
-            Some(false),
-            false,
-        );
-
-        assert_eq!(
-            super::classify_stale_active_turn_proof(&inflight, &snapshot, now_unix),
-            super::StaleActiveTurnProofClassification::QueueBlockedOrphan
-        );
-    }
-
-    /// #1456: explicit background placeholders are a visible status surface,
-    /// not disposable stale proof. Even if their inflight timestamp is old,
-    /// the fail-open classifier must preserve them instead of taking the
-    /// cleanup path that would cancel the owning session.
-    #[test]
-    fn stale_active_turn_proof_preserves_explicit_background_status() {
-        let provider = ProviderKind::Codex;
-        let channel_id = 900_000_000_000_911u64;
-        let now_unix = chrono::Utc::now().timestamp();
-        let stale_at = local_at_offset(
-            now_unix,
-            -(super::super::super::inflight::INFLIGHT_STALENESS_THRESHOLD_SECS as i64) - 5,
-        );
-        let mut inflight = inflight_with_updated_at(&provider, channel_id, &stale_at);
-        inflight.long_running_placeholder_active = true;
-        let snapshot = watcher_snapshot(
-            &provider,
-            channel_id,
-            inflight.user_msg_id,
-            false,
-            Some(false),
-            false,
-        );
-
-        assert_eq!(
-            super::classify_stale_active_turn_proof(&inflight, &snapshot, now_unix),
-            super::StaleActiveTurnProofClassification::ExplicitBackgroundStatus
-        );
-    }
-}
-
-/// #3009 — pure-logic tests for merged-placeholder reuse. These exercise the
-/// `pick_reusable_merged_placeholder` decision (which card a merged follow-up
-/// re-keys) without needing a live serenity ctx.
-#[cfg(test)]
-mod merged_placeholder_reuse_pure_tests {
-    use super::*;
-    use poise::serenity_prelude::MessageId;
-    use std::collections::HashMap;
-
-    #[test]
-    fn merged_followup_reuses_prior_head_card() {
-        // Queue head started as msg A (owns card CA); follow-up B merged in,
-        // so the head's source ids are [A, B] and the head message_id == B.
-        let a = MessageId::new(800_000_000_000_001);
-        let b = MessageId::new(800_000_000_000_002);
-        let card_a = MessageId::new(700_000_000_000_001);
-        let map: HashMap<MessageId, MessageId> = [(a, card_a)].into_iter().collect();
-
-        let picked = pick_reusable_merged_placeholder(&[a, b], b, |id| map.get(&id).copied());
-
-        // The follow-up B must reuse A's existing card (no new card).
-        assert_eq!(picked, Some((a, card_a)));
-    }
-
-    #[test]
-    fn second_merged_followup_reuses_rekeyed_card() {
-        // After B merged and re-keyed CA under B, a third follow-up C merges:
-        // source ids [A, B, C], only B currently owns the card.
-        let a = MessageId::new(800_000_000_000_001);
-        let b = MessageId::new(800_000_000_000_002);
-        let c = MessageId::new(800_000_000_000_003);
-        let card = MessageId::new(700_000_000_000_001);
-        let map: HashMap<MessageId, MessageId> = [(b, card)].into_iter().collect();
-
-        let picked = pick_reusable_merged_placeholder(&[a, b, c], c, |id| map.get(&id).copied());
-
-        assert_eq!(picked, Some((b, card)));
-    }
-
-    #[test]
-    fn no_prior_card_falls_back_to_fresh_post() {
-        // The head's earlier POST failed, so no source id owns a card yet.
-        let a = MessageId::new(800_000_000_000_001);
-        let b = MessageId::new(800_000_000_000_002);
-        let map: HashMap<MessageId, MessageId> = HashMap::new();
-
-        let picked = pick_reusable_merged_placeholder(&[a, b], b, |id| map.get(&id).copied());
-
-        assert_eq!(picked, None);
-    }
-
-    #[test]
-    fn own_id_card_is_never_reused() {
-        // Defensive: even if the follow-up id somehow already mapped to a card,
-        // it must be excluded so we never "reuse" the card we are about to
-        // create / treat the new head as its own prior.
-        let a = MessageId::new(800_000_000_000_001);
-        let b = MessageId::new(800_000_000_000_002);
-        let card_b = MessageId::new(700_000_000_000_002);
-        let map: HashMap<MessageId, MessageId> = [(b, card_b)].into_iter().collect();
-
-        let picked = pick_reusable_merged_placeholder(&[a, b], b, |id| map.get(&id).copied());
-
-        assert_eq!(picked, None);
-    }
-}
-
-/// #3082 part A pure decision tests for `pick_channel_queued_placeholder` — the
-/// "one card per active turn" coalescing that closes the gaps merge cannot
-/// (different author, reply-boundary, identical re-send outside the dedup
-/// window). These don't need a live serenity ctx.
-#[cfg(test)]
-mod channel_queued_placeholder_pure_tests {
-    use super::*;
-    use poise::serenity_prelude::MessageId;
-    use std::collections::HashSet;
-
-    fn ids(items: &[MessageId]) -> HashSet<MessageId> {
-        items.iter().copied().collect()
-    }
-
-    #[test]
-    fn different_author_reuses_existing_channel_card() {
-        // Author A queued first (card CA, still in queue). Author B's message —
-        // which does NOT merge with A — must reuse CA, not POST a second card.
-        let a = MessageId::new(900_000_000_000_001);
-        let b = MessageId::new(900_000_000_000_002);
-        let card_a = MessageId::new(710_000_000_000_001);
-        let cards = vec![(a, card_a)];
-        let queued = ids(&[a, b]);
-
-        let picked = pick_channel_queued_placeholder(&cards, b, &queued);
-        assert_eq!(picked, Some((a, card_a)), "B must coalesce onto A's card");
-    }
-
-    #[test]
-    fn reply_boundary_message_reuses_existing_card() {
-        // Same author, but the new message has a reply boundary so merge was
-        // skipped. Still must coalesce onto the single channel card.
-        let a = MessageId::new(900_000_000_000_010);
-        let b = MessageId::new(900_000_000_000_011);
-        let card_a = MessageId::new(710_000_000_000_010);
-        let cards = vec![(a, card_a)];
-        let queued = ids(&[a, b]);
-
-        let picked = pick_channel_queued_placeholder(&cards, b, &queued);
-        assert_eq!(picked, Some((a, card_a)));
-    }
-
-    #[test]
-    fn identical_resend_outside_dedup_window_reuses_card() {
-        // The identical-resend dedup only fires within 10s; a later identical
-        // re-send is enqueued as a new head but must still coalesce.
-        let a = MessageId::new(900_000_000_000_020);
-        let b = MessageId::new(900_000_000_000_021);
-        let card_a = MessageId::new(710_000_000_000_020);
-        let cards = vec![(a, card_a)];
-        let queued = ids(&[a, b]);
-
-        let picked = pick_channel_queued_placeholder(&cards, b, &queued);
-        assert_eq!(picked, Some((a, card_a)));
-    }
-
-    #[test]
-    fn prefers_live_queue_owner_over_stale_card() {
-        // Two cards exist on the channel: a stale one (owner drained) and the
-        // live waiting head. The live head's card must win so we re-key the
-        // canonical card; the stale one is left for normal cleanup.
-        let stale = MessageId::new(900_000_000_000_030);
-        let live = MessageId::new(900_000_000_000_031);
-        let new = MessageId::new(900_000_000_000_032);
-        let card_stale = MessageId::new(710_000_000_000_030);
-        let card_live = MessageId::new(710_000_000_000_031);
-        // Stale first so the loop would hit it before the live one — preference
-        // logic must still return the live card.
-        let cards = vec![(stale, card_stale), (live, card_live)];
-        let queued = ids(&[live, new]);
-
-        let picked = pick_channel_queued_placeholder(&cards, new, &queued);
-        assert_eq!(picked, Some((live, card_live)));
-    }
-
-    #[test]
-    fn falls_back_to_any_card_when_no_live_owner() {
-        // No card owner is still in the queue (all drained) but a visible card
-        // lingers — coalesce onto it rather than stacking a new one.
-        let stale = MessageId::new(900_000_000_000_040);
-        let new = MessageId::new(900_000_000_000_041);
-        let card_stale = MessageId::new(710_000_000_000_040);
-        let cards = vec![(stale, card_stale)];
-        let queued = ids(&[new]);
-
-        let picked = pick_channel_queued_placeholder(&cards, new, &queued);
-        assert_eq!(picked, Some((stale, card_stale)));
-    }
-
-    #[test]
-    fn no_existing_card_posts_fresh() {
-        // Nothing on the channel yet → None → caller posts the first card.
-        let new = MessageId::new(900_000_000_000_050);
-        let cards: Vec<(MessageId, MessageId)> = Vec::new();
-        let queued = ids(&[new]);
-
-        assert_eq!(pick_channel_queued_placeholder(&cards, new, &queued), None);
-    }
-
-    #[test]
-    fn own_card_is_never_self_reused() {
-        // Defensive: a card already keyed to the new head must be ignored so we
-        // never "reuse" the card we are about to create for it.
-        let new = MessageId::new(900_000_000_000_060);
-        let card_new = MessageId::new(710_000_000_000_060);
-        let cards = vec![(new, card_new)];
-        let queued = ids(&[new]);
-
-        assert_eq!(pick_channel_queued_placeholder(&cards, new, &queued), None);
-    }
-
-    #[test]
-    fn three_queued_messages_collapse_to_one_card() {
-        // Acceptance: queueing 3 messages leaves exactly one visible card.
-        // m1 posts CA. m2 reuses CA (re-keyed to m2). m3 reuses it again.
-        let m1 = MessageId::new(900_000_000_000_070);
-        let m2 = MessageId::new(900_000_000_000_071);
-        let m3 = MessageId::new(900_000_000_000_072);
-        let card = MessageId::new(710_000_000_000_070);
-
-        // m2: only m1 owns the card.
-        let picked_2 = pick_channel_queued_placeholder(&[(m1, card)], m2, &ids(&[m1, m2]));
-        assert_eq!(picked_2, Some((m1, card)), "m2 coalesces onto m1's card");
-
-        // After re-key, the card is owned by m2; m3 coalesces onto it.
-        let picked_3 = pick_channel_queued_placeholder(&[(m2, card)], m3, &ids(&[m1, m2, m3]));
-        assert_eq!(picked_3, Some((m2, card)), "m3 coalesces onto the one card");
-    }
-}
-
 #[cfg(test)]
 mod reply_context_tests {
-    use super::{
-        AttachmentReplyItem, format_attachment_reply_context, should_start_attachment_only_turn,
-        strip_leading_bot_mention,
-    };
+    use super::gate::{should_start_attachment_only_turn, strip_leading_bot_mention};
+    use super::{AttachmentReplyItem, format_attachment_reply_context};
 
     #[test]
     fn attachment_reply_context_keeps_canonical_message_id_and_all_files() {
